+++ tests/unit/Stack/Config/DockerSpec.hs
@@ -14,7 +14,7 @@ import           Test.Hspec ( Spec, describe, it, shouldBe )
 spec :: Spec
 spec = do
   describe "addDefaultTag" $ do
-    it "succeeds fails no snapshot resolver" $ addDefaultTag "foo/bar" Nothing Nothing `shouldBe` Nothing
+    it "succeeds fails no snapshot" $ addDefaultTag "foo/bar" Nothing Nothing `shouldBe` Nothing
     it "succeeds on LTS" $
       addDefaultTag
         "foo/bar"